requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
gunicorn==20.0.4
psycopg2-binary==2.9.7
pytest-django==4.5.2
pytz==2023.3.post1
sqlparse==0.4.4
